==> 01_scraping/06_find_relevant_offers.ipynb <==
import pandas as pd

data = pd.read_csv("05_offers_rake.csv")

import json

with open('thesaurus.json') as json_file:
    thesaurus = json.load(json_file)

data.info()

# Text-Ansatz

skill_found_text = []

for i in range(len(data)):
    
    counter = 0
    skills =[]
    
    for key in thesaurus.keys():
        if key in data["text"].iloc[i]:
            counter += 1
            skills.append(key)
        for j in range(len(thesaurus[key])):
            if " " + thesaurus[key][j] + " " in data["text"].iloc[i]:
                counter += 1
                skills.append(thesaurus[key][j])
        
    skill_found_text.append([counter, skills])

skill_found_text

numb = []
for skill in skill_found_text:
    numb.append(skill[0])
    
skills = []
for skill in skill_found_text:
    skills.append(skill[1])

data["number_of_skills_text"] = numb
data["skills_found_text"] = skills

comparison = data[["text", "number_of_skills_text", "skills_found_text"]]
comparison

comparison[comparison["number_of_skills_text"]  > 0]

# Keyphrases-Ansatz

data["keyphrases"].iloc[1000]

skill_found_keyphr = []

for i in range(len(data)):
    
    counter = 0
    skills =[]
    
    for key in thesaurus.keys():
        for j in range(len(thesaurus[key])):
            if "'" + thesaurus[key][j] + "'" in data["keyphrases"].iloc[i]:
                counter += 1
                skills.append(thesaurus[key][j])
        
    skill_found_keyphr.append([counter, skills])

numb = []
for skill in skill_found_keyphr:
    numb.append(skill[0])
    
skills = []
for skill in skill_found_keyphr:
    skills.append(skill[1])

data["number_of_skills_keyphr"] = numb
data["skills_found_keyphr"] = skills

comparison2 = data[["text", "keyphrases", "number_of_skills_text", "skills_found_text", "number_of_skills_keyphr", "skills_found_keyphr"]]
comparison2

comparison2[comparison2["number_of_skills_keyphr"]  > 0]

# Kombination

combi = comparison2[(comparison2["number_of_skills_keyphr"]  > 0 )| (comparison2["number_of_skills_text"] > 0)]
combi

combi[(combi["number_of_skills_text"] != 0) & (combi["number_of_skills_keyphr"] != 0)]

combi[combi["skills_found_text"] == combi["skills_found_keyphr"]]

# Listen umwandeln

Quelle: https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173

combi["keyphrases"] = combi["keyphrases"].apply(eval)

data["keyphrases"] = data["keyphrases"].apply(eval)

combi

def clean_alt_list(list_):
    list_ = list_.replace(', ', '","')
    list_ = list_.replace('[', '["')
    list_ = list_.replace(']', '"]')
    return list_

#combi["skills_found_text"] = combi["skills_found_text"].apply(clean_alt_list)
#combi["skills_found_keyphr"] = combi["skills_found_keyphr"].apply(clean_alt_list)

Dies ist gar nicht nötig, es sind bereits Lists.

def to_1D(series):
    return pd.Series([x for _list in series for x in _list])

to_1D(combi["skills_found_text"]).value_counts()

to_1D(combi["skills_found_keyphr"]).value_counts()

keyphrases = to_1D(combi["keyphrases"]).value_counts().rename_axis('unique_values').reset_index(name='counts')

keyphrases.to_csv('keyphrases.csv', index=False)
